Changed user-model relation - several users for one model

+++ routes/model.js
@@ -9,7 +9,7 @@ const validateObjectId = require("../middleware/validateObjectId");
 const {
   exists,
   checkIfFileExistsAsync,
-  removeFileOrDirectoryAsync
+  removeFileOrDirectoryAsync,
 } = require("../utilities/utilities");
 const hasUser = require("../middleware/auth/hasUser");
 const isAdmin = require("../middleware/auth/isAdmin");
@@ -35,7 +35,7 @@ router.get("/:userId", [hasUser, isAdmin], async (req, res) => {
 
   //Finding model based on user id
   let models = await Model.find({
-    user: req.params.userId
+    user: req.params.userId,
   });
 
   //Returning all models payload
@@ -60,7 +60,7 @@ router.get(
     //Finding model based on user id and model id
     let model = await Model.findOne({
       _id: req.params.id,
-      user: req.params.userId
+      user: req.params.userId,
     });
     if (!exists(model)) return res.status(404).send("Model or user not found");
 
@@ -83,7 +83,7 @@ router.post(
     if (!exists(user)) return res.status("404").send("User not found...");
 
     //Generating new model
-    let model = new Model({ name: req.body.name, user: req.params.userId });
+    let model = new Model({ name: req.body.name, user: [req.params.userId] });
 
     await model.save();
 
@@ -111,27 +111,44 @@ router.delete(
     //Finding model to delete
     let model = await Model.findOne({
       _id: req.params.id,
-      user: req.params.userId
+      user: req.params.userId,
     });
 
     //Returning if model not found
     if (!exists(model)) return res.status("404").send("Model not found...");
 
+    //Removing userId from model
+    model.user = model.user.filter(
+      (element) => element.toString() !== user._id.toString()
+    );
+
     //Generating payload to return of deleted model
-    let payloadToReturn = await model.getPayload();
+    let payloadToReturn = null;
 
-    //removing model
-    await Model.deleteOne({ _id: req.params.id });
+    //Checking if model should be deleted or edited
+    if (model.user.length > 0) {
+      //Editing model only
+      await model.save();
 
-    //removing model file if exists
-    let modelFilePath = Project.getModelFilePath(user, model);
-    let modelFileExists = await checkIfFileExistsAsync(modelFilePath);
-    if (modelFileExists) await removeFileOrDirectoryAsync(modelFilePath);
+      payloadToReturn = await model.getPayload();
+    } else {
+      //Deleting model
+      payloadToReturn = await model.getPayload();
 
-    //removing model file if exists
-    let modelIOSFilePath = Project.getModelIOSFilePath(user, model);
-    let modelIOSFileExists = await checkIfFileExistsAsync(modelIOSFilePath);
-    if (modelIOSFileExists) await removeFileOrDirectoryAsync(modelIOSFilePath);
+      //removing model
+      await Model.deleteOne({ _id: req.params.id });
+
+      //removing model file if exists
+      let modelFilePath = Project.getModelFilePath(model);
+      let modelFileExists = await checkIfFileExistsAsync(modelFilePath);
+      if (modelFileExists) await removeFileOrDirectoryAsync(modelFilePath);
+
+      //removing model file if exists
+      let modelIOSFilePath = Project.getModelIOSFilePath(model);
+      let modelIOSFileExists = await checkIfFileExistsAsync(modelIOSFilePath);
+      if (modelIOSFileExists)
+        await removeFileOrDirectoryAsync(modelIOSFilePath);
+    }
 
     logger.action(`User ${req.user.email} deleted model ${model._id}`);
 
@@ -154,17 +171,23 @@ router.put(
     //Finding model to edit
     let model = await Model.findOne({
       _id: req.params.id,
-      user: req.params.userId
+      user: req.params.userId,
     });
 
     //Returning if model not found
     if (!exists(model)) return res.status("404").send("Model not found...");
 
-    //If user exists in body and is different than user of model - return 400
-    if (exists(req.body.user) && req.body.user !== model.user.toString())
-      return res.status("400").send("Model of user cannot be changed");
+    //Editing users if exist
+    if (exists(req.body.user)) {
+      //If user exists in body - checking if users exists for given array body
+      for (let userId of req.body.user) {
+        newUser = await User.findOne({ _id: userId });
+        if (!exists(newUser))
+          return res.status("404").send("User in user property not found ...");
+      }
 
-    //Editing model
+      model.user = req.body.user;
+    }
 
     //Editing name if exists
     if (exists(req.body.name)) model.name = req.body.name;
